minor cleanup and refactoring
