+++ warp-core/src/main/scala/com/workday/warp/inject/WarpGuicer.scala
@@ -3,11 +3,11 @@ package com.workday.warp.inject
 import java.lang.reflect.Constructor
 
 import com.google.inject.{AbstractModule, Guice, Injector}
+import com.workday.warp.TestId
 import com.workday.warp.collectors.AbstractMeasurementCollectionController
 import com.workday.warp.common.{PropertyEntry, WarpPropertyLike}
-import com.workday.warp.junit.TestIdConverters._
+import com.workday.warp.TestIdImplicits._
 import com.workday.warp.inject.modules.{DefaultWarpModule, HasWarpBindings}
-import com.workday.warp.junit.HasTestId
 import com.workday.warp.persistence.Tag
 import org.junit.jupiter.api.TestInfo
 import org.pmw.tinylog.Logger
@@ -93,10 +93,10 @@ object WarpGuicer {
     * @param tags tags to use for this test.
     * @return a measurement controller.
     */
-  def getController(hasTestId: HasTestId, tags: Iterable[Tag]): AbstractMeasurementCollectionController = {
+  def getController(hasTestId: TestId, tags: Iterable[Tag]): AbstractMeasurementCollectionController = {
     this.getController(hasTestId.testId, tags)
   }
-  def getController(hasTestId: HasTestId): AbstractMeasurementCollectionController = this.getController(hasTestId, Nil)
+  def getController(hasTestId: TestId): AbstractMeasurementCollectionController = this.getController(hasTestId, Nil)
 
 
   /**
