+++ warp-core/src/test/scala/com/workday/warp/arbiters/ResponseTimeArbiterSpec.scala
@@ -5,7 +5,7 @@ import java.time.Instant
 
 import com.workday.telemetron.annotation.Required
 import com.workday.warp.common.spec.WarpJUnitSpec
-import com.workday.warp.junit.TestIdConverters._
+import com.workday.warp.TestIdImplicits._
 import com.workday.warp.junit.UnitTest
 import com.workday.warp.persistence.CorePersistenceAware
 import com.workday.warp.persistence.TablesLike.TestExecutionRowLike
