+++ warp-core/src/test/scala/com/workday/warp/arbiters/ZScoreArbiterSpec.scala
@@ -6,7 +6,7 @@ import java.util.UUID
 import com.workday.telemetron.RequirementViolationException
 import com.workday.warp.common.annotation.ZScoreRequirement
 import com.workday.warp.common.spec.WarpJUnitSpec
-import com.workday.warp.junit.TestIdConverters._
+import com.workday.warp.TestIdImplicits._
 import com.workday.warp.junit.UnitTest
 import com.workday.warp.persistence.CorePersistenceAware
 import com.workday.warp.persistence.TablesLike.TestExecutionRowLike
